FHist

FHist.Hist1D — Method

Hist1D(array, edges::AbstractRange; kwgs...)
Hist1D(array, edges::AbstractVector; error_mode=:sqrt)

Create a Hist1D with given bin edges and vlaues from array. Weight for each value is assumed to be 1.

FHist.Hist1D — Method

Hist1D(array, wgts::AbstractWeights, edges::AbstractRange, ; kwgs...)
Hist1D(array, wgts::AbstractWeights, edges::AbstractVector; error_mode=:sqrt)

Create a Hist1D with given bin edges and vlaues from array. wgts should have the same size as array.

FHist.Hist1D — Method

Hist1D(A::AbstractVector{T}; nbins::Integer=StatsBase.sturges(length(A)), error_mode=:sqrt) where T
Hist1D(A::AbstractVector{T}, wgts::AbstractWeights; nbins::Integer=StatsBase.sturges(length(A)), error_mode=:sqrt) where T

Automatically determine number of bins based on Sturges algo.

FHist.Hist1D — Method

Hist1D(h::Histogram{T, 1, E}; error_mode=:sqrt) where {T,E}

Convert an existing 1D StatsBase.Histogram to a Hist1D. Adds error according to error_mode.

FHist.Hist1D — Method

Hist1D(elT::Type{T}=Float64; binedges) where {T}

Initialize an empty histogram with bin content typed as T and bin edges. To be used with push!

Base.push! — Method

push!(h::Hist1D, val::Real, wgt::Real=one{T})

Adding one value at a time into histogram. Remember to call update_error! after if you need errors.

FHist.update_error! — Function

update_error!(h::Hist1D, error_fun = sqrt_err)

Update the error (up and down) of a histogram according to a specific error_mode. Remember to call this function after updaing a histogram with push!() in a loop.