FHist.Hist1D — Method

Hist1D(array, edges::AbstractRange)
Hist1D(array, edges::AbstractVector)

Create a Hist1D with given bin edges and vlaues from array. Weight for each value is assumed to be 1.

FHist.Hist1D — Method

Hist1D(array, wgts::AbstractWeights, edges::AbstractRange)
Hist1D(array, wgts::AbstractWeights, edges::AbstractVector)

Create a Hist1D with given bin edges and vlaues from array. wgts should have the same size as array.

FHist.Hist1D — Method

Hist1D(A::AbstractVector{T}; nbins::Integer=StatsBase.sturges(length(A))) where T
Hist1D(A::AbstractVector{T}, wgts::AbstractWeights; nbins::Integer=StatsBase.sturges(length(A))) where T

Automatically determine number of bins based on Sturges algo.

FHist.Hist1D — Method

Hist1D(elT::Type{T}=Float64; binedges) where {T}

Initialize an empty histogram with bin content typed as T and bin edges. To be used with push!

FHist.Hist2D — Method

Hist2D(tuple, wgts::AbstractWeights, edges::Tuple{AbstractRange,AbstractRange})
Hist2D(tuple, wgts::AbstractWeights, edges::Tuple{AbstractVector,AbstractVector})

Create a Hist2D with given bin edges and values from a 2-tuple of arrays of x, y values. wgts should have the same size as elements of tuple.

FHist.Hist2D — Method

Hist2D(tuple, edges::Tuple{AbstractRange,AbstractRange})
Hist2D(tuple, edges::Tuple{AbstractVector,AbstractVector})

Create a Hist2D with given bin edges and values from a 2-tuple of arrays of x, y values. Weight for each value is assumed to be 1.

FHist.Hist2D — Method

Hist2D(A::AbstractVector{T}; nbins::Tuple{Integer,Integer}) where T
Hist2D(A::AbstractVector{T}, wgts::AbstractWeights; nbins::Tuple{Integer,Integer}) where T

Automatically determine number of bins based on Sturges algo.

FHist.Hist2D — Method

Hist2D(elT::Type{T}=Float64; binedges) where {T}

Initialize an empty histogram with bin content typed as T and bin edges. To be used with push!

Base.empty! — Method

empty!(h::Hist1D)

Resets a histogram's bin counts and sumw2.

Base.empty! — Method

empty!(h::Hist2D)

Resets a histogram's bin counts and sumw2.

Base.push! — Method

unsafe_push!(h::Hist1D, val::Real, wgt::Real=1)
push!(h::Hist1D, val::Real, wgt::Real=1)

Adding one value at a time into histogram. sumw2 (sum of weights^2) accumulates wgt^2 with a default weight of 1. unsafe_push! is a faster version of push! that is not thread-safe.

N.B. To append multiple values at once, use broadcasting via push!.(h, [-3.0, -2.9, -2.8]) or push!.(h, [-3.0, -2.9, -2.8], 2.0)

Base.push! — Method

unsafe_push!(h::Hist2D, valx::Real, valy::Real, wgt::Real=1)
push!(h::Hist2D, valx::Real, valy::Real, wgt::Real=1)

Adding one value at a time into histogram. sumw2 (sum of weights^2) accumulates wgt^2 with a default weight of 1. unsafe_push! is a faster version of push! that is not thread-safe.

FHist.bincenters — Method

bincenters(h::Hist1D)

Get the bin centers of a histogram.

FHist.bincenters — Method

bincenters(h::Hist2D)

Get a 2-tuple of the bin centers of a histogram.

FHist.bincounts — Method

bincounts(h::Hist1D)

Get the bin counts of a histogram.

FHist.bincounts — Method

bincounts(h::Hist2D)

Get the bin counts of a histogram.

FHist.binedges — Method

binedges(h::Hist1D)

Get the bin edges of a histogram.

FHist.binedges — Method

binedges(h::Hist2D)

Get a 2-tuple of the bin edges of a histogram.

FHist.cumulative — Method

cumulative(h::Hist1D; forward=true)

Create a cumulative histogram. If forward, start summing from the left.

FHist.hists_to_bars — Method

hists_to_bars(hist1ds)

Given a vector of Hist1D, return edges (xs), heights (ys), and grps (for grouping) that is useful for plotting stacked histogram.

FHist.integral — Method

integral(h::Hist1D)

Get the integral a histogram.

FHist.integral — Method

integral(h::Hist2D)

Get the integral a histogram.

FHist.lookup — Method

function lookup(h::Hist1D, x)

For given x-axis value x, find the corresponding bin and return the bin content. If a value is out of the histogram range, return missing.

FHist.lookup — Method

function lookup(h::Hist2D, x, y)

For given x-axis and y-axis value x, y, find the corresponding bin and return the bin content. If a value is out of the histogram range, return missing.

FHist.nbins — Method

nbins(h::Hist1D)

Get the number of bins of a histogram.

FHist.nbins — Method

nbins(h::Hist2D)

Get a 2-tuple of the number of x and y bins of a histogram.

FHist.profile — Function

profile(h::Hist2D, axis::Symbol=:x)
profile(axis::Symbol=:x) = h::Hist2D -> profile(h, axis)

Returns the axis-profile of the 2D histogram by calculating the weighted mean over the other axis. profile(h, :x) will return a Hist1D with the y-axis edges of h.

FHist.project — Function

project(h::Hist2D, axis::Symbol=:x)
project(axis::Symbol=:x) = h::Hist2D -> project(h, axis)

Computes the :x (:y) axis projection of the 2D histogram by summing over the y (x) axis. Returns a Hist1D.

FHist.rebin — Function

rebin(h::Hist1D, n::Int=1)
rebin(n::Int) = h::Hist1D -> rebin(h, n)

Merges n consecutive bins into one. The returned histogram will have nbins(h)/n bins.

FHist.rebin — Function

rebin(h::Hist2D, nx::Int=1, ny::Int=nx)
rebin(nx::Int, ny::Int) = h::Hist2D -> rebin(h, nx, ny)

Merges nx (ny) consecutive bins into one along the x (y) axis by summing.

FHist.restrict — Function

restrict(h::Hist1D, low=-Inf, high=Inf)
restrict(low=-Inf, high=Inf) = h::Hist1D -> restrict(h, low, high)

Returns a new histogram with a restricted x-axis. restrict(h, 0, 3) (or h |> restrict(0, 3)) will return a slice of h where the bin centers are in [0, 3] (inclusive).

FHist.sample — Method

sample(h::Hist1D, n::Int=1)

Sample a histogram's with weights equal to bin count, n times. The sampled values are the bins' lower edges.

FHist.transpose — Method

transpose(h::Hist2D)

Reverses the x and y axes.

LinearAlgebra.normalize — Method

normalize(h::Hist1D)

Create a normalized histogram via division by integral(h).

LinearAlgebra.normalize — Method

normalize(h::Hist2D)

Create a normalized histogram via division by integral(h).

Statistics.mean — Method

Statistics.mean(h::Hist1D)
Statistics.std(h::Hist1D)
Statistics.median(h::Hist1D)
Statistics.quantile(h::Hist1D, p)

Compute statistical quantities based on the bin centers weighted by the bin counts.