FHist.Hist1D — Method

Hist1D(array, edges::AbstractRange)
Hist1D(array, edges::AbstractVector)

Create a Hist1D with given bin edges and vlaues from array. Weight for each value is assumed to be 1.

FHist.Hist1D — Method

Hist1D(array, wgts::AbstractWeights, edges::AbstractRange)
Hist1D(array, wgts::AbstractWeights, edges::AbstractVector)

Create a Hist1D with given bin edges and vlaues from array. wgts should have the same size as array.

FHist.Hist1D — Method

Hist1D(A::AbstractVector{T}; nbins::Integer=StatsBase.sturges(length(A))) where T
Hist1D(A::AbstractVector{T}, wgts::AbstractWeights; nbins::Integer=StatsBase.sturges(length(A))) where T

Automatically determine number of bins based on Sturges algo.

FHist.Hist1D — Method

Hist1D(elT::Type{T}=Float64; binedges) where {T}

Initialize an empty histogram with bin content typed as T and bin edges. To be used with push!

Base.push! — Method

push!(h::Hist1D, val::Real)
push!(h::Hist1D, val::Real, wgt::Real=one{T})

Adding one value at a time into histogram. If wgt is supplied , this operation will accumulate sumw2 (sum of weights^2) in the Hist automatically.

FHist.sample — Method

sample(h::Hist1D)
sample(h::Hist1D, n::Int)

Sample a histogram's with weights equal to bin count, one or n times. The returned sample value will be one of the bin's left edge.